Refactor downloading code to its own files
