+++ frontend/src/store/fetch.js
@@ -1,0 +1,13 @@
+import {ref} from 'vue';
+
+const loading = ref(0);
+
+export async function fetchJSON(url, options) {
+	try {
+		loading.value += 1;
+		const response = await fetch(url, options);
+		return await response.json();
+	} finally {
+		loading.value -= 1;
+	}
+}
