+++ frontend/src/views/AddDatasetView.vue
@@ -1,9 +1,17 @@
 <script setup>
 import {ref, reactive, computed, watch, onMounted} from 'vue';
-import { Interval } from '../interval.js';
 import { formatSize } from '../format.js';
 import VueSelect from 'vue-select';
 import {DownloadCloudIcon} from 'vue3-feather';
+import DownloadPopup from '../components/DownloadPopup.vue';
+import {fetchJSON} from '../store/fetch.js';
+import {
+	download,
+	isDownloading,
+	fetchDownloadableDatasets,
+	fetchSourceLanguages,
+	fetchTargetLanguages
+} from '../store/downloads.js';
 
 import 'vue-select/dist/vue-select.css';
 
@@ -32,8 +40,6 @@ const SORT_ORDER_OPTIONS = [
 ];
 
 const sortOrder = ref(SORT_ORDER_OPTIONS[0]);
-
-const loading = ref(0);
 
 // Per language all target languages
 const languages = new Map();
@@ -112,7 +118,7 @@ const datasets = computed(() => {
 		const list = ref([]);
 		cache.set(key, list);
 		// Fetches actual list async, but the cache entry is available immediately.
-		fetchDatasets(key).then(datasets => list.value = datasets);
+		fetchDownloadableDatasets(key).then(datasets => list.value = datasets);
 	}
 
 	// cache contains refs, so this computed() is called again once the data
@@ -149,34 +155,12 @@ const datasets = computed(() => {
 	return datasets;
 });
 
-const downloads = reactive({});
-
 onMounted(async () => {
 	fetchSourceLanguages().then(languages => {
 		srcLangs.value = languages;
 	})
-
-	fetchDownloads().then(list => {
-		Object.assign(downloads, castDownloadListToMap(list))
-	})
 })
 
-let downloadUpdateInterval = new Interval(1000, async () => {
-	const list = await fetchDownloads();
-	Object.assign(downloads, castDownloadListToMap(list));
-})
-
-// Watch list of downloads to re-evaluate whether we need to do updating using
-// the downloadUpdateInterval. Stop updating once there's nothing active
-// anymore. The changes caused by downloadSelection() will re-trigger this
-// watch expression and enable the interval again.
-watch(downloads, (downloads) => {
-	const activeStates = new Set(['pending', 'downloading']);
-	if (Object.values(downloads).some(download => activeStates.has(download.state)))
-		downloadUpdateInterval.restart();
-	else
-		downloadUpdateInterval.stop();
-});
 
 function assignList(current, update, key = 'id') {
 	const updates = Object.fromEntries(update.map(entry => [entry[key], entry]));
@@ -184,52 +168,6 @@ function assignList(current, update, key = 'id') {
 		if (current[i][key] in updates)
 			Object.assign(current[i], updates[current[i][key]]);
 	return current;
-}
-
-function castDownloadListToMap(list) {
-	return Object.fromEntries(list.map(download => [download.entry.id, download]));
-}
-
-function download(dataset) {
-	requestDownloadSelection([dataset]).then(update => {
-		Object.assign(downloads, castDownloadListToMap(update));
-	});
-}
-
-async function fetchJSON(url, options) {
-	try {
-		loading.value += 1;
-		const response = await fetch(url, options);
-		return await response.json();
-	} finally {
-		loading.value -= 1;
-	}
-}
-
-async function fetchSourceLanguages() {
-	return await fetchJSON('/api/download/languages/');
-}
-
-async function fetchTargetLanguages(sourceLanguage) {
-	return await fetchJSON(`/api/download/languages/${encodeURIComponent(sourceLanguage)}`);
-}
-
-async function fetchDatasets(key) {
-	return await fetchJSON(`/api/download/by-language/${encodeURIComponent(key)}`);
-}
-
-async function fetchDownloads() {
-	return await fetchJSON(`/api/download/downloads/`);
-}
-
-async function requestDownloadSelection(datasets) {
-	return await fetchJSON('/api/download/downloads/', {
-		method: 'POST',
-		headers: {
-			'Content-Type': 'application/json'
-		},
-		body: JSON.stringify(datasets.map(({id}) => ({id})))
-	});
 }
 
 const countFormat = new Intl.NumberFormat();
@@ -272,7 +210,7 @@ const countFormat = new Intl.NumberFormat();
 			<div class="dataset" v-for="dataset in datasets" :key="dataset.id" :id="`did-${dataset.id}`">
 				<div class="dataset-name">
 					<h3 class="dataset-title"><a :href="`https://opus.nlpl.eu/${dataset.corpus}-${dataset.version}.php`" target="_blank">{{ dataset.corpus }}</a></h3>
-					<button class="download-dataset-button" @click="download(dataset)" :disabled="dataset.id in downloads || 'paths' in dataset">
+					<button class="download-dataset-button" @click="download(dataset)" :disabled="isDownloading(dataset) || 'paths' in dataset">
 						Download
 						<DownloadCloudIcon class="download-icon"/>
 					</button>
@@ -290,12 +228,7 @@ const countFormat = new Intl.NumberFormat();
 			</div>
 		</div>
 		<Teleport to=".navbar">
-			<details class="downloads-popup">
-				<summary><h2>Downloads</h2></summary>
-				<ul>
-					<li v-for="download in downloads" :key="download.entry.id">{{ download.entry.corpus }} <em>{{ download.state }}</em></li>
-				</ul>
-			</details>
+			<DownloadPopup/>
 		</Teleport>
 	</div>
 </template>
